# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Order API",
        "description": "API Documentation for Order Management",
        "termsOfService": "https://www.example.com/terms",
        "contact": {
            "name": "Support Team",
            "email": "support@example.com",
            "url": "https://www.example.com/contact"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://interviews.test/api",
            "description": "Development API"
        }
    ],
    "paths": {
...
        "/cart": {
            "get": {
                "tags": [
                    "cart"
                ],
                "summary": "View cart items",
                "description": "Retrieves all items currently in the cart",
                "operationId": "viewCart",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Cart is empty"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "product_code": {
                                                        "type": "string",
                                                        "example": "Pineapple_Inc_KE_8aztdcizacca"
                                                    },
                                                    "product_name": {
                                                        "type": "string",
                                                        "example": "Floveme Power Bank 2S 20000mAh"
                                                    },
                                                    "product_quantity": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "product_price": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 2000
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch cart"
                    }
                }
            },
            "post": {
                "tags": [
                    "cart"
                ],
                "summary": "Add item to cart",
                "description": "Adds a product to the Redis cart",
                "operationId": "addToCart",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "product_code": {
                                        "type": "string",
                                        "example": "Pineapple_Inc_KE_8aztdcizacca"
                                    },
                                    "product_name": {
                                        "type": "string",
                                        "example": "Floveme Power Bank 2S 20000mAh"
                                    },
                                    "product_quantity": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "product_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 2000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Item added to cart",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Item added to cart"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to add item to cart",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to add item to cart"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "cart"
                ],
                "summary": "Clear the cart",
                "description": "Removes all items currently in the cart",
                "operationId": "clearCart",
                "responses": {
                    "200": {
                        "description": "Cart cleared successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cart cleared"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to clear the cart"
                    }
                }
            }
        },
        "/orders": {
            "get": {
                "tags": [
                    "orders"
                ],
                "summary": "Retrieve all orders",
                "description": "Returns a list of all orders",
                "operationId": "getAllOrders",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "order_id": {
                                                "type": "string",
                                                "example": "RONDO-ABCDEFGH12"
                                            },
                                            "items": {
                                                "type": "string",
                                                "example": "[{'product_code': '...', ...}]"
                                            },
                                            "total": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 2000
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "orders"
                ],
                "summary": "Place an order",
                "description": "Creates a new order based on items in the Redis cart",
                "operationId": "placeOrder",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "product_code": {
                                                    "type": "string",
                                                    "example": "Pineapple_Inc_KE_8aztdcizacca"
                                                },
                                                "product_name": {
                                                    "type": "string",
                                                    "example": "Floveme Power Bank 2S 20000mAh"
                                                },
                                                "product_quantity": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "product_price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 2000
                                                },
                                                "product_total": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 2000
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order placed successfully"
                    },
                    "400": {
                        "description": "Cart is empty",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cart is empty"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}